requirements.txt
streamlit>=1.32.0
tensorflow>=2.16.0
numpy>=1.26.0
Pillow>=10.0.0
streamlit-paste-button
